I don't remember what this is because I left it for months.

+++ libc/string/memcmp.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <string.h>
 
 int memcmp(const void *string1, const void *string2, size_t n)
@@ -5,7 +6,7 @@ int memcmp(const void *string1, const void *string2, size_t n)
     const unsigned char *s1 = (const unsigned char *) string1;
     const unsigned char *s2 = (const unsigned char *) string2;
 
-    for(int i = 0; i < n; i++)
+    for(size_t i = 0; i < n; i++)
     {
         if(s1[i] - s2[i] != 0)
         {
